requirements.txt
bottlenose==1.1.8
lxml==4.2.1
nose==1.3.7
numpy==1.14.3
pandas==0.22.0
python-amazon-simple-product-api==2.2.11
python-dateutil==2.7.3
pytz==2018.4
six==1.11.0
